requirements.txt
playwright>=1.41.0
python-dotenv>=1.0.0

# Installation instructions:
# 1. Install Python dependencies: pip install -r requirements.txt
# 2. Install Playwright browsers: playwright install chromium
